numerous bug fixes

+++ blacknova/scheduler.php
@@ -70,37 +70,40 @@ else
 {
  
   $sched_res = $db->Execute("SELECT * FROM $dbtables[scheduler]");
-  while (!$sched_res->EOF)
+  if($sched_res)
   {
-    $event = $sched_res->fields;
-    $multiplier = ($sched_ticks / $event[ticks_full]) + ($event[ticks_left] / $event[ticks_full]);
-    $multiplier = (int) $multiplier;
-    $ticks_left = ($sched_ticks + $event[ticks_left]) % $event[ticks_full];
+     while (!$sched_res->EOF)
+     {
+       $event = $sched_res->fields;
+       $multiplier = ($sched_ticks / $event[ticks_full]) + ($event[ticks_left] / $event[ticks_full]);
+       $multiplier = (int) $multiplier;
+       $ticks_left = ($sched_ticks + $event[ticks_left]) % $event[ticks_full];
 
-    if($event[loop] == 'N')
-    {
-      if($multiplier > $event[spawn])
-        $multiplier = $event[spawn];
+       if($event[loop] == 'N')
+       {
+         if($multiplier > $event[spawn])
+           $multiplier = $event[spawn];
       
-      if($event[spawn] - $multiplier == 0)
-        $db->Execute("DELETE FROM $dbtables[scheduler] WHERE sched_id=$event[sched_id]");
-      else
-        $db->Execute("UPDATE $dbtables[scheduler] SET ticks_left=$ticks_left, spawn=spawn-$multiplier WHERE sched_id=$event[sched_id]");
-    }
-    else
-      $db->Execute("UPDATE $dbtables[scheduler] SET ticks_left=$ticks_left WHERE sched_id=$event[sched_id]");
+         if($event[spawn] - $multiplier == 0)
+           $db->Execute("DELETE FROM $dbtables[scheduler] WHERE sched_id=$event[sched_id]");
+         else
+           $db->Execute("UPDATE $dbtables[scheduler] SET ticks_left=$ticks_left, spawn=spawn-$multiplier WHERE sched_id=$event[sched_id]");
+       }
+       else
+         $db->Execute("UPDATE $dbtables[scheduler] SET ticks_left=$ticks_left WHERE sched_id=$event[sched_id]");
   
-    $sched_var_id = $event[sched_id];
-    $sched_var_extrainfo = $event[extra_info];
+       $sched_var_id = $event[sched_id];
+       $sched_var_extrainfo = $event[extra_info];
 
-    $sched_i = 0;
-    while($sched_i < $multiplier)
-    {
-      include("$event[file]");
-      $sched_i++;
-    }
-    $sched_res->MoveNext();
-  }
+       $sched_i = 0;
+       while($sched_i < $multiplier)
+       {
+         include("$event[file]");
+         $sched_i++;
+       }
+       $sched_res->MoveNext();
+     }
+   }
   
   include("footer.php");
   if($sched_type == 1) ob_end_clean();
